backend/requirements.txt
# requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
